move building use and building creator into one file, small fixes
